refactor(parsers): move comfyui parsing logic from tool.py into parsers/comfyui.py
